MFlow

Description

Non monadic low level support stuff for the MFlow application server. it implements an scheduler of queued Processable messages that are served according with the source identification and the verb invoked. Ths scheduler executed the appropriate workflow (using the workflow package) the workflow may send additional messages to the source, identified by a Token . The computation state is logged and can be recovered.

The message communication is trough polimorphic, monoidal queues. There is no asumption about message codification, so instantiations of this scheduler for many different infrastructures is possible. MFlow.Hack is an instantiation for the Hack interface in a Web context.

MFlow.Wai is a instantiation for the WAI interface.

In order to manage resources, the serving process may die after a timeout. as well as the logged state, usually, after a longer timeout .

All these details are hidden in the monad of MFlow.Forms that provides an higuer level interface. Altroug fragments streaming sendFragment sendEndFragment are only provided at this level.

stateless and transient serving processes are possible. stateless are request-response with no intermediate messaging dialog. transient processes have no persistent state, so they restart anew after a timeout or a crash.

Documentation

type Params = [(String, String)]

data Req

Constructors

forall a . (Processable a, Typeable a) => Req a

data Resp

Constructors

Fragm HttpData

EndFragm HttpData

Resp HttpData

type Workflow m = WF Stat m

data HttpData

Constructors

HttpData Params [Cookie] ByteString

Error WFErrors ByteString

class Processable a where

Methods

pwfname :: a -> String

puser :: a -> String

pind :: a -> String

getParams :: a -> Params

class ToHttpData a where

Methods

toHttpData :: a -> HttpData

data Token

Constructors

Token

Fields

twfname :: String

tuser :: String

tind :: String

q :: MVar Req

qr :: MVar Resp

getToken :: Processable a => a -> IO Token

type ProcList = WorkflowList IO Token ()

List of (wfname, workflow) pairs, to be scheduled depending on the message's pwfname

flushRec :: Token -> IO ()

receive :: Typeable a => Token -> IO a

receiveReq :: Token -> IO Req

receiveReqTimeout :: Int -> Integer -> Token -> IO Req

send :: ToHttpData a => Token -> a -> IO ()

send a complete response

sendFlush :: ToHttpData a => Token -> a -> IO ()

sendFragment :: ToHttpData a => Token -> a -> IO ()

send a response fragment. Useful for streaming. the last packet must sent trough send

sendEndFragment :: ToHttpData a => Token -> a -> IO ()

msgScheduler :: (Typeable a, Processable a) => a -> IO (HttpData, ThreadId)

addMessageFlows :: [(String, Token -> Workflow IO ())] -> IO ()

getMessageFlows :: IO (Map String (Token -> Workflow IO ()))

transient :: (Token -> IO ()) -> Token -> Workflow IO ()

to add a monadic computation that send and receive messages, but does not store its state in permanent storage. The process once stopped, will restart anew

stateless :: ToHttpData b => (Params -> IO b) -> Token -> Workflow IO ()

anonymous :: [Char]

noScript :: [Char]

hlog :: Handle

addTokenToList :: Token -> IO ()

deleteTokenInList :: Token -> IO ()

btag :: String -> Attribs -> ByteString -> ByteString

Writes a XML tag in a ByteString. It is the most basic form of formatting. For more sophisticated formatting , use Text.XHtml or HSP.

bhtml :: Attribs -> ByteString -> ByteString

 bhtml ats v= btag "html" ats v

bbody :: Attribs -> ByteString -> ByteString

 bbody ats v= btag "body" ats v

type Attribs = [(String, String)]