MFlow.Forms.Admin

Documentation

adminLoop :: IO ()

A small console interpreter with some commands:

sync

Synchronize the cache with persistent storage (see syncCache)

flush

Flush the cache

end

Synchronize and exit

abort

Exit. Do not synchronize

on exception, for example Control-c, it sync and exit. it used as the last statement of the main procedure.

addAdminWF :: IO ()

Install the admin flow in the list of flows handled by HackMessageFlow this gives access to an administrator page. It is necessary to create an admin user with setAdminUser.