MFlow.Hack

Documentation

module MFlow.Cookies

module MFlow

hackMessageFlow :: [(String, Token -> Workflow IO ())] -> Env -> IO Response

An instance of the abstract MFlow scheduler to the Hack interface it accept the list of processes being scheduled and return a hack handler

Example:

main= do

putStrLn $ options messageFlows
   run 80 $ hackMessageFlow  messageFlows
   where
   messageFlows=  [("main",  runFlow flowname )
                  ,("hello", stateless statelesproc)
                  ,("trans", transient $ runflow transientflow]
   options msgs= "in the browser choose\n\n" ++
     concat [ http://server/++ i ++ n | (i,_) <- msgs]